changed all increment to biginc and int to big int, as it required by laravel 5.8

+++ database/migrations/2019_06_10_061850_create_events_table.php
@@ -14,11 +14,11 @@ class CreateEventsTable extends Migration
     public function up()
     {
         Schema::create('events', function (Blueprint $table) {
-                $table->increments('id');
+                $table->bigIncrements('id');
                 $table->string('name');
                 $table->mediumText('whatsapp_link')->nullable();
 
-                $table->integer('leader_id')->unsigned();
+                $table->bigInteger('leader_id')->unsigned();
                 $table->foreign('leader_id')->references('id')->on('users');
 
                 $table->mediumText('description');
