+++ database/migrations/2019_06_10_065350_create_points_table.php
@@ -14,13 +14,13 @@ class CreatePointsTable extends Migration
     public function up()
     {
         Schema::create('points', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->integer('value');
 
-            $table->integer('approved_by_admin_id')->unsigned();
+            $table->bigInteger('approved_by_admin_id')->unsigned();
             $table->foreign('approved_by_admin_id')->references('id')->on('users');
 
-            $table->integer('task_id')->unsigned();
+            $table->bigInteger('task_id')->unsigned();
             $table->foreign('task_id')->references('id')->on('tasks');
 
             $table->timestamps();
