+++ database/migrations/2019_06_10_072216_users_events.php
@@ -14,13 +14,14 @@ class UsersEvents extends Migration
     public function up()
     {
             Schema::create('users_events', function (Blueprint $table) {
-                $table->increments('id');
+                $table->bigIncrements('id');
 
-                $table->integer('user_id')->unsigned();
+                $table->bigInteger('user_id')->unsigned();
                 $table->foreign('user_id')->references('id')->on('users');
 
-                $table->integer('event_id')->unsigned();
+                $table->bigInteger('event_id')->unsigned();
                 $table->foreign('event_id')->references('id')->on('events');
+
                 $table->timestamps();
             });
